--- biome.json
@@ -1,10 +1,0 @@
-{
-	"$schema": "https://biomejs.dev/schemas/2.3.10/schema.json",
-	"assist": { "actions": { "source": { "organizeImports": "on" } } },
-	"linter": {
-		"enabled": true,
-		"rules": {
-			"recommended": true
-		}
-	}
-}
