+++ subscript_superscript.json
@@ -1,112 +1,112 @@
 {
-    "superscript": {
-        "a": "ᵃ",
-        "b": "ᵇ",
-        "c": "ᶜ",
-        "d": "ᵈ",
-        "e": "ᵉ",
-        "f": "ᶠ",
-        "g": "ᵍ",
-        "h": "ʰ",
-        "i": "ⁱ",
-        "j": "ʲ",
-        "k": "ᵏ",
-        "l": "ˡ",
-        "m": "ᵐ",
-        "n": "ⁿ",
-        "o": "ᵒ",
-        "p": "ᵖ",
-        "r": "ʳ",
-        "s": "ˢ",
-        "t": "ᵗ",
-        "u": "ᵘ",
-        "v": "ᵛ",
-        "w": "ʷ",
-        "x": "ˣ",
-        "y": "ʸ",
-        "z": "ᶻ",
-        "A": "ᴬ",
-        "B": "ᴮ",
-        "D": "ᴰ",
-        "E": "ᴱ",
-        "G": "ᴳ",
-        "H": "ᴴ",
-        "I": "ᴵ",
-        "J": "ᴶ",
-        "K": "ᴷ",
-        "L": "ᴸ",
-        "M": "ᴹ",
-        "N": "ᴺ",
-        "O": "ᴼ",
-        "P": "ᴾ",
-        "R": "ᴿ",
-        "T": "ᵀ",
-        "U": "ᵁ",
-        "V": "ⱽ",
-        "W": "ᵂ",
-        "0": "⁰",
-        "1": "¹",
-        "2": "²",
-        "3": "³",
-        "4": "⁴",
-        "5": "⁵",
-        "6": "⁶",
-        "7": "⁷",
-        "8": "⁸",
-        "9": "⁹",
-        "alpha": "ᵅ",
-        "beta": "ᵝ",
-        "gamma": "ᵞ",
-        "delta": "ᵟ",
-        "epsilon": "ᵋ",
-        "theta": "ᶿ",
-        "iota": "ᶥ",
-        "phi_2": "ᶲ",
-        "phi": "ᵠ",
-        "xi": "ᵡ",
-        "+": "⁺",
-        "-": "⁻",
-        "=": "⁼",
-        "(": "⁽",
-        ")": "⁾"
-    },
-    "subscript": {
-        "a": "ₐ",
-        "e": "ₑ",
-        "h": "ₕ",
-        "i": "ᵢ",
-        "j": "ⱼ",
-        "k": "ₖ",
-        "l": "ₗ",
-        "m": "ₘ",
-        "n": "ₙ",
-        "o": "ₒ",
-        "p": "ₚ",
-        "r": "ᵣ",
-        "s": "ₛ",
-        "t": "ₜ",
-        "u": "ᵤ",
-        "v": "ᵥ",
-        "x": "ₓ",
-        "0": "₀",
-        "1": "₁",
-        "2": "₂",
-        "3": "₃",
-        "4": "₄",
-        "5": "₅",
-        "6": "₆",
-        "7": "₇",
-        "8": "₈",
-        "9": "₉",
-        "beta": "ᵦ",
-        "gama": "ᵧ",
-        "rho": "ᵨ",
-        "phi": "ᵩ",
-        "xi": "ᵪ",
-        "+": "₊",
-        "-": "₋",
-        "=": "₌",
-        "(": "₍",
-        ")": "₎"
-    }
+  "superscript": {
+    "a": "ᵃ",
+    "b": "ᵇ",
+    "c": "ᶜ",
+    "d": "ᵈ",
+    "e": "ᵉ",
+    "f": "ᶠ",
+    "g": "ᵍ",
+    "h": "ʰ",
+    "i": "ⁱ",
+    "j": "ʲ",
+    "k": "ᵏ",
+    "l": "ˡ",
+    "m": "ᵐ",
+    "n": "ⁿ",
+    "o": "ᵒ",
+    "p": "ᵖ",
+    "r": "ʳ",
+    "s": "ˢ",
+    "t": "ᵗ",
+    "u": "ᵘ",
+    "v": "ᵛ",
+    "w": "ʷ",
+    "x": "ˣ",
+    "y": "ʸ",
+    "z": "ᶻ",
+    "A": "ᴬ",
+    "B": "ᴮ",
+    "D": "ᴰ",
+    "E": "ᴱ",
+    "G": "ᴳ",
+    "H": "ᴴ",
+    "I": "ᴵ",
+    "J": "ᴶ",
+    "K": "ᴷ",
+    "L": "ᴸ",
+    "M": "ᴹ",
+    "N": "ᴺ",
+    "O": "ᴼ",
+    "P": "ᴾ",
+    "R": "ᴿ",
+    "T": "ᵀ",
+    "U": "ᵁ",
+    "V": "ⱽ",
+    "W": "ᵂ",
+    "0": "⁰",
+    "1": "¹",
+    "2": "²",
+    "3": "³",
+    "4": "⁴",
+    "5": "⁵",
+    "6": "⁶",
+    "7": "⁷",
+    "8": "⁸",
+    "9": "⁹",
+    "alpha": "ᵅ",
+    "beta": "ᵝ",
+    "gamma": "ᵞ",
+    "delta": "ᵟ",
+    "epsilon": "ᵋ",
+    "theta": "ᶿ",
+    "iota": "ᶥ",
+    "phi_2": "ᶲ",
+    "phi": "ᵠ",
+    "xi": "ᵡ",
+    "+": "⁺",
+    "-": "⁻",
+    "=": "⁼",
+    "(": "⁽",
+    ")": "⁾"
+  },
+  "subscript": {
+    "a": "ₐ",
+    "e": "ₑ",
+    "h": "ₕ",
+    "i": "ᵢ",
+    "j": "ⱼ",
+    "k": "ₖ",
+    "l": "ₗ",
+    "m": "ₘ",
+    "n": "ₙ",
+    "o": "ₒ",
+    "p": "ₚ",
+    "r": "ᵣ",
+    "s": "ₛ",
+    "t": "ₜ",
+    "u": "ᵤ",
+    "v": "ᵥ",
+    "x": "ₓ",
+    "0": "₀",
+    "1": "₁",
+    "2": "₂",
+    "3": "₃",
+    "4": "₄",
+    "5": "₅",
+    "6": "₆",
+    "7": "₇",
+    "8": "₈",
+    "9": "₉",
+    "beta": "ᵦ",
+    "gama": "ᵧ",
+    "rho": "ᵨ",
+    "phi": "ᵩ",
+    "xi": "ᵪ",
+    "+": "₊",
+    "-": "₋",
+    "=": "₌",
+    "(": "₍",
+    ")": "₎"
+  }
 }
